requirements.txt
asttokens==2.2.1
backcall==0.2.0
decorator==5.1.1
entrypoints==0.3
executing==1.2.0
flake8==3.7.8
ipython==8.12.2
jedi==0.19.0
matplotlib-inline==0.1.6
mccabe==0.6.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.5.0
pycryptodome==3.18.0
pyflakes==2.1.1
Pygments==2.16.1
six==1.16.0
stack-data==0.6.2
traitlets==5.9.0
typing-extensions==4.7.1
virtualenv==16.7.7
wcwidth==0.2.6
